+++ source/common/quic/quic_filter_manager_connection_impl.cc
@@ -11,7 +11,7 @@ namespace Quic {
 
 QuicFilterManagerConnectionImpl::QuicFilterManagerConnectionImpl(
     QuicNetworkConnection& connection, const quic::QuicConnectionId& connection_id,
-    Event::Dispatcher& dispatcher, uint64_t send_buffer_limit,
+    Event::Dispatcher& dispatcher, uint32_t send_buffer_limit,
     std::shared_ptr<QuicSslConnectionInfo>&& info)
     // Using this for purpose other than logging is not safe. Because QUIC connection id can be
     // 18 bytes, so there might be collision when it's hashed to 8 bytes.
